Don't copy namespace when deserialising functions
Doing this meant that functions inside the request namespace each got a copy of
the namespace at whatever progress deserialisation was happening which then
stayed static, so if the main function performed an import it wouldn't see it,
and if it was the first deserialised function it would not see the rest, etc.

Now we don't copy at all which should be fine.